Add uptime check to prevent early power off
Raspberry Pi's don't come with an RTC and therefore time is all
over the place on startup. This causes a jump which triggers the
poweroff code. This commit adds a check to diff uptime against the
user-set timeout and to wait until that has passed.

This is probably a very inelegant way of doing this as it assumes
that /proc/uptime exists and the user has set a timeout. We could
just force it to wait for a minute to allow ntp to sync.